requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.3
black==24.4.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.3
colorama==0.4.6
configue==5.0.0
contourpy==1.2.1
coqpit==0.0.17
crcmod==1.7
cryptography==42.0.7
cycler==0.12.1
datasets==2.20.0
decord==0.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
einops==0.8.0
email_validator==2.1.1
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.2
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gguf==0.6.0
gradio==4.29.0
gradio_client==0.16.1
grpcio==1.65.4
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
llvmlite==0.42.0
lm-format-enforcer==0.10.3
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
mteb==1.13.1
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
openai==1.30.4
opencv-python==4.10.0.84
openpyxl==3.1.3
orjson==3.10.3
oss2==2.18.5
outlines==0.0.43
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pdf2image==1.17.0
peft==0.11.1
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
polars==1.4.1
portalocker==2.8.2
prometheus_client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
protobuf==4.25.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytrec-eval-terrier==0.5.6
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.1
ruff==0.4.6
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
sse-starlette==2.1.0
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.6.0
timm==0.9.10
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
transformers==4.44.0
transformers-stream-generator==0.0.5
triton==3.0.0
trl==0.8.6
typer==0.12.3
typing_extensions==4.12.0
tyro==0.8.4
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
vllm==0.5.4
vllm-flash-attn==2.6.1
vllm_nccl_cu12==2.18.1.0.4.0
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
wheel==0.43.0
xformers==0.0.27.post2
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zhipuai==2.1.4.20230806
zipp==3.19.0
zstandard==0.22.0